Add support for params generation
This changes the top level cli to have “index” and “params” as the two main commands to either generate or dump
