+++ generator/Cargo.toml
@@ -12,11 +12,13 @@ readme = "README.md"
 anyhow = {version = "1", features = ["backtrace"]}
 clap = {version = "4", features = ["derive"]}
 geojson = {version = "0.24", features = ["geo-types"]}
+helium-proto = { git = "https://github.com/helium/proto", branch="master" }
+serde = {version = "1", features = ["derive"]}
+serde_json = "1"
 flate2 = "1"
 h3ron = {version = "0.16"}
 geo-types = "*"
 byteorder = "1.4"
-micro-timer = "0"
 rayon = "1.6"
 log = "0"
 simple_logger = "4"
